+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Get all pages when listing CodeArtifact versions
+
 ## [0.4.2] - 2024-03-26
 
 ### Changed